requirements.txt
diffusers
torch
pinecone
fashion-clip
python-dotenv
boto3
fastapi
uvicorn
pydantic
Pillow
transformers
beautifulsoup4
